chore: moves survey follow ups to community
